Update variables gathering to use vars/varnames lookups
Directly accessing the internal `vars` variable in ansible is not
supported and the same functionality is available with the vars
and varnames lookups.

Note the use of *query() which converts the list returned from
query() into a set of positional arguments for the subsequent
query, as query does not accept a list.

Change-Id: I74931fed262a44b6c3b410f7120ce443123a2268