Create relative links to root instead of absolute
We don't really need to create an absolute links for root/intermediate
certificates since original file is placed at the same place. In case
of storing PKI dir in git these links are considered as  out-of-bound
which results in sanity checks failure.
Also we're eliminating risk of broken links in case when PKI directory
is being moved on the filesystem

Change-Id: Ia7f4f4726bf4d73c6c0071bef52fd6eb58d662a3