Only create certificate destination directories once.
If installing several certs or keys to a target the destination
directory creation task is run once for each list item even if
it is the same directory every time.

This patch filters the list of targets and reduces the list of
target directories to a unique list, reducing the number of loop
iterations required.

Change-Id: I7432463dfa067afed2f46874119378bfdf96639c