Separate backend-specific logic to improve extensibility
Although the role is intended to support multiple backends, in practice
it was tightly coupled with the standalone backend. Several tasks in the
"generic" parts were specific to the standalone case, and the
distinction between global and backend-specific variables was unclear.

This patch improves the separation between generic and backend-specific
logic, making it easier to add and maintain additional backends in the
future.

Change-Id: I9124c85099e1338c2cb2dcef971475eaf3a5860d
Signed-off-by: Damian Dabrowski <damian.dabrowski@cleura.com>