Rename pki_method variable to pki_backend
It should be clearly communicated that this variable only defines a
default backend for generated certificates and can be overridden on
per-certificate basis.
This patch also changes naming convention from "method" to "backend"
because it sounds like a more universal approach that would be easier to
understand for the users.
Backwards compatbility is supported and should be kept for a few
releases.

Change-Id: I5a929c59212ca1cd61d85cb7c3ee4a6903164b27
Signed-off-by: Damian Dabrowski <damian.dabrowski@cleura.com>