Change the format of 'san' parameter in `pki_certificates` variable
The format of `san` parameter was changed from a string to the dictonary
of lists.

Example:

```
san:
  dns:
    - "{{ ansible_facts['hostname'] }}"
  ip:
    - "{{ management_address }}"
  uri: []
  other: []
```

It will be extremely useful when adding new backends to PKI role.
For example, hashi_vault requires to define dns/ip/uri and other sans in
completely different variables, while in standalone plugin we need to
combine all of them in a string.
But generating a string in required format from the above dict is way
easier than generating a dict in an expected format based on a string.

Change-Id: I2716e06207e31d34cd250e77ed7b555418b99492
Signed-off-by: Damian Dabrowski <damian.dabrowski@cleura.com>