Allow certificates to be installed by specifying them by name
Previously the full path to the certificate on the PKI host
was required to be in the `src` parameter to the pki role.
This means that internal knowledge about the implementation
of the PKI role was needed in the code using the role.

This change allows the certificate to be installed to be
specified by name, in exactly the same way that the name can
be specificed when the certificate is created with the standalone
backend.

The `src` parameter is now reserved for when the user wants to
pass in a path to their own certificate, and install directly
from that instead.

Change-Id: Ib38cc344ac42783af5c4f7a0d74a9bb310b879ff
Signed-off-by: Jonathan Rosser <jonathan.rosser@rd.bbc.co.uk>