Fix creation of certs signed by selfsigned issuers
Certificates in ansible-role-pki can be signed by:
1. intermediate certificate(it has provider=='ownca')
2. directly by a root certificate(it has provider=='selfsigned')

In scenario 1., ansible-role-pki creates a *-chain.crt file containing a
CA bundle containing the full chain of trust[1].

In scenario 2., the *-chain.crt file is not created because it doesn't
make any sense. There is no chain of trust because certificate is signed
directly by a root CA.

In patch [2], the `ca_bundle` cert format was implemented(the approach
was slightly changed in [3] later on).
So, when a new certificate is created, ansible-role-pki needs to be
aware of the full chain of trust for this certificate in order to create
a file containing 'ca_bundle'.
Currently, the previously mentioned
`"{{ pki_dir ~ '/roots/' ~ cert.signed_by ~ '/certs/' ~ cert.signed_by ~ '-chain.crt' }}"`
file is used for this purpose.
But because this file is created only for "ownca" issuers, this approach
is not suitable for "selfsigned" issuers, where this file just doesn't
exist.

As a solution, this patch makes sure that *-chain.crt file always
exists, even for a standalone CA roots.

[1] https://opendev.org/openstack/ansible-role-pki/src/branch/master/tasks/standalone/create_ca.yml#L157-L166
[2] https://review.opendev.org/c/openstack/ansible-role-pki/+/954628
[2] https://review.opendev.org/c/openstack/ansible-role-pki/+/956576

Change-Id: I3baf488fa25f6cb27f2ac330f9eef4c9203c0498
Signed-off-by: Damian Dabrowski <damian.dabrowski@cleura.com>